Adds the posibility to define different setup pages for Blaze Dashboard
