+++ apps/blaze-dashboard/src/pages/controller.js
@@ -1,6 +1,7 @@
+import config from '@automattic/calypso-config';
 import BlazeSetup from './setup/main';
 
 export const setup = ( context, next ) => {
-	context.primary = <BlazeSetup />;
+	context.primary = <BlazeSetup setupInfo={ config( 'need_setup' ) } />;
 	next();
 };
