+++ apps/blaze-dashboard/src/pages/setup/main.jsx
@@ -7,14 +7,21 @@ import MainWrapper from 'calypso/my-sites/promote-post-i2/components/main-wrappe
 import PostsListBanner from 'calypso/my-sites/promote-post-i2/components/posts-list-banner';
 import WooBanner from 'calypso/my-sites/promote-post-i2/components/woo-banner';
 import { getAdvertisingDashboardPath } from 'calypso/my-sites/promote-post-i2/utils';
-import { useSelector } from 'calypso/state';
-import { getSelectedSite } from 'calypso/state/ui/selectors';
 import GenericHeader from '../../components/generic-header';
+import BlazeDisabled from './components/blaze-disabled';
 
-export default function BlazeSetup() {
-	const selectedSiteData = useSelector( getSelectedSite );
-	const adminUrl = selectedSiteData?.options?.admin_url;
+const renderSetupComponent = ( setupInfo ) => {
+	// Compatibility with previous Jetpack's version
+	const reason = typeof setupInfo === 'boolean' ? 'blaze_disabled' : setupInfo;
 
+	switch ( reason ) {
+		case 'blaze_disabled':
+		default:
+			return <BlazeDisabled />;
+	}
+};
+
+export default function BlazeSetup( { setupInfo } ) {
 	const isWooBlaze = config.isEnabled( 'is_running_in_woo_site' );
 	const translate = useTranslate();
 
@@ -43,36 +50,7 @@ export default function BlazeSetup() {
 
 			<div className="promote-post-i2__aux-wrapper">
 				<div className="empty-promotion-list__container promote-post-i2__setup-container">
-					<h3 className="empty-promotion-list__title wp-brand-font">
-						{ translate( 'Set up Blaze and start promoting your store' ) }
-					</h3>
-					<p className="empty-promotion-list__body">
-						{ translate(
-							"You're on the brink of unleashing the advertising potential of your store. To get started, follow these simple steps below:"
-						) }
-					</p>
-
-					<ol className="promote-post-i2__active-steps">
-						<li>
-							{ translate( 'Go to the {{a}}traffic section{{/a}} of the Jetpack settings page.', {
-								components: {
-									a: (
-										<a
-											href={ `${ adminUrl }/admin.php?page=jetpack#/traffic` }
-											target="_blank"
-											rel="noreferrer"
-										/>
-									),
-								},
-							} ) }
-						</li>
-						<li>{ translate( 'Scroll down to the Blaze section and toggle it to active.' ) }</li>
-					</ol>
-					<p className="empty-promotion-list__body">
-						{ translate(
-							"After completing these steps, you're all set! Return here, refresh the page and start promoting your store with Blaze!"
-						) }
-					</p>
+					{ renderSetupComponent( setupInfo ) }
 				</div>
 			</div>
 		</MainWrapper>
